Specify more granular Batik dependencies
Batik-all is a strange artifact. It's POM declares dependencies on all the sub-JARs, but its JAR has all of the sub-jars repackaged. This results in two JARs with the package "org.apache.batik.transcoder" being added to consuming applications.

The Ant build does not use batik-all, so the Maven and Gradle builds should not either.